requirements.txt
# TFST Carrier Portal - Python Dependencies

# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Session==0.5.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Salesforce Integration
simple-salesforce==1.12.6
requests==2.31.0

# Firebase Integration
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0

# AWS S3 Integration
boto3==1.34.34
botocore==1.34.34

# File Processing
pandas==2.1.4
python-csv==1.0
openpyxl==3.1.2

# Security & Authentication
cryptography==41.0.8
PyJWT==2.8.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30
Werkzeug==3.0.1

# Web Server (Production)
gunicorn==21.2.0
waitress==2.1.2

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0

# Monitoring & Logging
sentry-sdk[flask]==1.39.2

# Background Jobs (Optional)
celery==5.3.4
redis==5.0.1

# API Documentation
flasgger==0.9.7.1

# Rate Limiting
Flask-Limiter==3.5.0

# Environment Variables
python-decouple==3.8

# JSON Processing
jsonschema==4.20.0

# HTTP Client
httpx==0.25.2

# Image Processing (for document handling)
Pillow==10.1.0

# Caching
Flask-Caching==2.1.0

# WSGI
uWSGI==2.0.23